Move PremiumComponent to billing ownership
